2° BackEnd/requirements.txt
pipenv
django-seed
bcrypt
djangorestframework-simplejwt
djangorestframework
django-cors-headers
psycopg2
python-decouple
django-extensions
